riscv: Add initial semihosting support
This adds a new library libsemihost that implements syscalls using
semihosting syscalls. It is intended to be used instead of libgloss when
semihosting support is required. A new semihost.specs file has been
added which will link with -lsemihost instead -lgloss.

lseek SEEK_CUR is not currently supported and using it will set errno to
EINVAL and return -1. To support it requires adding an additional
mechanism to libsemihost to keep track of each open file and its
position.